r flytter EgenvurderingKort inn i Sluttkort for kss
